# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/products": {
      "post": {
        "operationId": "ProductsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "409": {
            "description": "Product code already exists"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create product",
        "tags": [
          "Products"
        ]
      },
      "get": {
        "operationId": "ProductsController_list",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Free-text search across name/code/category name.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "required": false,
            "in": "query",
            "description": "Filter by name (contains, case-insensitive).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "code",
            "required": false,
            "in": "query",
            "description": "Filter by code (contains, case-insensitive).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryName",
            "required": false,
            "in": "query",
            "description": "Filter by category name (contains, case-insensitive).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "description": "Filter by category id (26 chars).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Max items per page (1..100).",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Cursor for pagination (id < cursor).",
            "schema": {
              "minLength": 26,
              "maxLength": 26,
              "type": "string"
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "If provided, filters by isActive.",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Products list"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List products",
        "tags": [
          "Products"
        ]
      }
    },
    "/api/v1/products/attribute-definitions": {
      "post": {
        "operationId": "ProductsController_createAttributeDefinition",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductAttributeDefinitionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Attribute definition created"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "409": {
            "description": "Attribute key already exists for this category"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create product attribute definition",
        "tags": [
          "Products"
        ]
      },
      "get": {
        "operationId": "ProductsController_listAttributeDefinitions",
        "parameters": [
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "description": "Optional category id (26 chars) to filter by.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Attribute definitions list"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List product attribute definitions",
        "tags": [
          "Products"
        ]
      }
    },
    "/api/v1/products/attribute-definitions/{id}": {
      "patch": {
        "operationId": "ProductsController_updateAttributeDefinition",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Definition id (26 chars)",
            "schema": {
              "minLength": 26,
              "maxLength": 26,
              "example": "01J1QZQ0VQ8J7TQH0YV3A1BCDE",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductAttributeDefinitionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Attribute definition updated"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "404": {
            "description": "Attribute definition not found"
          },
          "409": {
            "description": "Attribute key already exists for this category"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update product attribute definition",
        "tags": [
          "Products"
        ]
      },
      "delete": {
        "operationId": "ProductsController_removeAttributeDefinition",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Definition id (26 chars)",
            "schema": {
              "minLength": 26,
              "maxLength": 26,
              "example": "01J1QZQ0VQ8J7TQH0YV3A1BCDE",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Attribute definition deleted"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "404": {
            "description": "Attribute definition not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete product attribute definition",
        "tags": [
          "Products"
        ]
      }
    },
    "/api/v1/products/{id}": {
      "get": {
        "operationId": "ProductsController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Product id (26 chars)",
            "schema": {
              "minLength": 26,
              "maxLength": 26,
              "example": "01J1QZQ0VQ8J7TQH0YV3A1BCDE",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product detail"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "404": {
            "description": "Product not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get product by id",
        "tags": [
          "Products"
        ]
      },
      "patch": {
        "operationId": "ProductsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Product id (26 chars)",
            "schema": {
              "minLength": 26,
              "maxLength": 26,
              "example": "01J1QZQ0VQ8J7TQH0YV3A1BCDE",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product updated"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "404": {
            "description": "Product not found"
          },
          "409": {
            "description": "Product code already exists"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update product",
        "tags": [
          "Products"
        ]
      },
      "delete": {
        "operationId": "ProductsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Product id (26 chars)",
            "schema": {
              "minLength": 26,
              "maxLength": 26,
              "example": "01J1QZQ0VQ8J7TQH0YV3A1BCDE",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product deleted"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "404": {
            "description": "Product not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete product",
        "tags": [
          "Products"
        ]
      }
    },
...
  },
  "info": {
    "title": "Stock SaaS API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Send a valid JWT access token. This API also accepts the access token via cookie (see cookie auth)."
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken",
        "description": "Optional. If present, the API will read the access token from this cookie."
      }
    },
    "schemas": {
...
      "InitialStockEntryDto": {
        "type": "object",
        "properties": {
          "branchId": {
            "type": "string",
            "description": "Branch id (26 chars) to initialize stock for.",
            "example": "01J1QZQ0VQ8J7TQH0YV3A1BCDE",
            "minLength": 26,
            "maxLength": 26
          },
          "stockOnHand": {
            "type": "number",
            "description": "Initial stock on hand (>= 0).",
            "example": 10,
            "minimum": 0
          },
          "price": {
            "type": "number",
            "description": "Initial price (>= 0).",
            "example": 1000,
            "minimum": 0
          }
        },
        "required": [
          "branchId",
          "stockOnHand",
          "price"
        ]
      },
      "CreateProductDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Product code (unique per tenant).",
            "example": "SKU-001",
            "maxLength": 64
          },
          "name": {
            "type": "string",
            "description": "Product name.",
            "example": "Remera Lisa",
            "maxLength": 200
          },
          "categoryId": {
            "type": "string",
            "description": "Category id (26 chars).",
            "minLength": 26,
            "maxLength": 26
          },
          "description": {
            "type": "string",
            "description": "Description.",
            "maxLength": 5000
          },
          "isActive": {
            "type": "boolean",
            "description": "Soft active flag. Defaults to true.",
            "default": true
          },
          "attributes": {
            "type": "object",
            "description": "Product custom attributes (key/value). Validated against category definitions."
          },
          "initialStock": {
            "description": "Optional initial stock entries to initialize inventories.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InitialStockEntryDto"
            }
          }
        },
        "required": [
          "code",
          "name"
        ]
      },
      "CreateProductAttributeDefinitionDto": {
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "string",
            "description": "Category id (26 chars).",
            "minLength": 26,
            "maxLength": 26
          },
          "key": {
            "type": "string",
            "description": "Attribute key (lowercase letters, numbers, underscore).",
            "example": "talle",
            "maxLength": 64
          },
          "label": {
            "type": "string",
            "description": "Attribute label.",
            "example": "Talle",
            "maxLength": 120
          },
          "type": {
            "type": "string",
            "description": "Attribute type.",
            "enum": [
              "TEXT",
              "NUMBER",
              "BOOLEAN",
              "DATE",
              "ENUM"
            ]
          },
          "unit": {
            "type": "string",
            "description": "Optional unit.",
            "example": "cm",
            "maxLength": 24
          },
          "isRequired": {
            "type": "boolean",
            "description": "Whether the attribute is required."
          },
          "isVisibleInTable": {
            "type": "boolean",
            "description": "Whether the attribute is visible in product table."
          },
          "sortOrder": {
            "type": "number",
            "description": "Sort order (0..1000).",
            "minimum": 0,
            "maximum": 1000
          },
          "options": {
            "description": "Options for ENUM attributes. Required when type=ENUM.",
            "example": [
              "S",
              "M",
              "L"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "categoryId",
          "key",
          "label",
          "type"
        ]
      },
      "UpdateProductAttributeDefinitionDto": {
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "string",
            "description": "Move definition to another category (26 chars)."
          },
          "key": {
            "type": "string",
            "description": "Attribute key (lowercase letters, numbers, underscore).",
            "example": "talle",
            "maxLength": 64
          },
          "label": {
            "type": "string",
            "description": "Attribute label.",
            "example": "Talle",
            "maxLength": 120
          },
          "type": {
            "type": "string",
            "description": "Attribute type.",
            "enum": [
              "TEXT",
              "NUMBER",
              "BOOLEAN",
              "DATE",
              "ENUM"
            ]
          },
          "unit": {
            "type": "string",
            "description": "Optional unit.",
            "example": "cm",
            "maxLength": 24
          },
          "isRequired": {
            "type": "boolean",
            "description": "Whether the attribute is required."
          },
          "isVisibleInTable": {
            "type": "boolean",
            "description": "Whether the attribute is visible in product table."
          },
          "sortOrder": {
            "type": "number",
            "description": "Sort order (0..1000).",
            "minimum": 0,
            "maximum": 1000
          },
          "options": {
            "description": "Options for ENUM attributes. Required when type=ENUM.",
            "example": [
              "S",
              "M",
              "L"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "UpdateProductDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Product code.",
            "example": "SKU-001",
            "maxLength": 64
          },
          "name": {
            "type": "string",
            "description": "Product name.",
            "example": "Remera Lisa",
            "maxLength": 200
          },
          "categoryId": {
            "type": "object",
            "description": "Category id (26 chars). Use null to clear."
          },
          "description": {
            "type": "string",
            "description": "Description. Use empty string to clear.",
            "maxLength": 5000
          },
          "isActive": {
            "type": "boolean",
            "description": "Soft active flag."
          },
          "attributes": {
            "type": "object",
            "description": "Product custom attributes (key/value). Validated against category definitions."
          }
        }
      },
...
    }
  }
}